Ensure CaptureSessionState is always finalized.
For resource cleanup reasons we want to make sure that
CaptureSessionState is always finalized so that resources can be
released. This fixes an edge case where the object can be disconnected
before the CameraDevice is ever opened. In this situation the
onSessionFinalized call will not be invoked, but disconnect() will be.

In addition, the change also adds tests for this behavior as well as
the onSessionFinalized behavior in CaptureSessionState.

Bug: 251296842

Test: ./gradlew\
 :camera:camera-camera2-pipe:testDebugUnitTest\
 :camera:camera-camera2-pipe-testing:testDebugUnitTest\
 :camera:camera-camera2-pipe-integration:testDebugUnitTest

Change-Id: I92e706549bb7100c717858eaaac666ed595f0de9