Prevent transient MeteringRepeating and make UseCaseManager thread-safe
This CL fixes a scenario in which:

1. ImageCapture, MeteringRepeating use cases are attached.
2. ImageCapture use case is detached.
3. MeteringRepeating use case is left attached.
4. MeteringRepeating is quickly detached after.

During 3., it results in a meaningless CameraGraph being created.

This CL fixes this by checking whether we should detach
MeteringRepeating as well when use cases are detached. We also add a
simular fix to check whether MeteringRepeating should be attached when
use cases are attached.

In addition, we add a lock to make sure UseCaseManager is thread-safe.

Bug: 251296842
Test: CameraPipe unit tests
Change-Id: Icb25b37fba4b29c0ebcef6d6f3774ab93133f7d7