Prevent NoSuchElementException in NavHost
There is a scenario when you can cause your NavGraph to reset during
recomposition and Crossfade does not have access to a composing
destination.

It is rare that reseting the graph is actually the desired behavior, but
at the very least, we can avoid crashing.

RelNote: "`NavHost` will no longer cause a `NoSuchElementException` when
there is not destination available for the `Crossfade` to compose. It
will now just skip the composition."
Test: tested in sample app
Bug: 253299416

Change-Id: Ieb46e8bbcbcde6183a2ab07fba24b18756a4ae2f