Depend on latest stable lifecycle prebuilts in Compose
We previously had to add a ToT dependency on lifecycle-common-java8 in
Compose to resolve dependency conflicts where old and new versions both
had a DefaultLifecycleEventObserver classes. see aosp/1805059.

But now that a stable version of this dep has been released, we should
move from ToT to prebuilts to handle this error while we work on
updating the lifecycle dependencies in Compose.

Test: ./gradlew bOS
Change-Id: If004a0d91614ca2e70686884edaeedda69fad38f