Fix Camera2CameraControlDeviceTest
After aosp/2217381, Camera2CameraControl is not in the set of
UseCaseCameraControl. The UseCaseManager doesn't set the request
options to the camera since Camera2CameraControl is not in the
java.util.Set<UseCaseCameraControl>.

This CL adds the Camera2CameraControl to the Set<UseCaseCameraControl>,
to make the Camera2CameraControl can update the camera.

Bug: 254006542
Test: Camera2CameraControlDeviceTest
Change-Id: Icbe520b382a6972e85fdde2f045b69d1e6d23b74