Improve handling invalidations for large number of composers
Reduced the overhead of notifying a large number of composers
of invalidations by pre-collapsing the invalidation set as the
chagnes are received from the snapshot system  instead of
waiting until the next composition starts.

No new tests were added as the behavior should be identical
to the previous behavior; just faster in stress scenarios.

Test: ./gradlew :composer:r:r:tDUT
Change-Id: I706e5bce819685347e0245a6576f27d926b69385