Run 2 ToolbarMenuHostTest only on API 24 and above
When animations are enabled on API 23 the providedMenuItemSelected and
manuallyInflatedMenuItemSelected tests in the ToolbarMenuHostTest suite
fail due to a race condition between when the menu items are
available and when we can click on them. Ideally, there would be a wait
for us to wait until the transition is complete, but on API 23 the
underlying APIs are not available.

For tests, animations should always be off anyway, but disabling the
tests for devices below API 24 to ensure that if animations are on we
can avoid failing since we know this isn't a problem on real devices.

Test: disables tests
Bug: 238664379
Change-Id: I427e9d0269f074cf059a1be234e5b8260a58b03c