Placeholder - work around anti-aliasing issues with Brushes being used over the top of Painters.
If you draw with a brush over the top of a painter there is a possibility of the color from the painter bleeding through around the edges due to alpha blending of the different layers.
As a result we need to workaround this by merging the colors rather than layering them where we can.

Bug: 253667329
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Relnote: "We have updated the placeholder background drawing to merge colors rather than layer them were possible to reduce the risk of alpha blending of the different cropped layers from allowing underlying colors to bleed through at the placeholder background edges."
Change-Id: I2ea2612412ca1fb3cad61c98cfee71ce85bee99f