In ImageCapture, recreate ImagePipeline when SessionConfig is reset.
Currently in the new pipeline, when onError is received, the old Surface is reused. I suspect that this is causing the "Surface had no valid native window" issue on legacy devices.

In this CL, we update the new pipeline's behavior to match the old pipeline's, that is, recreating the the Surface when onError() is called.

Also: removed a obsolete test.

Bug: 255737797
Test: manual test and ./gradlew bOS
Change-Id: I2c36db90a06fd1c2e4235e55835c6983fc07e54b