Take min dimensions in DP for SplitPairRule/SplitPlaceholderRule
Before, we take minWidth, minSmallestWidth in pixels, which doesn't get
updated when the display density is changed. Now, we take them in DP, so
that we can convert it to pixels later with the up-to-date density.

Also, refactor to take minWidth, minSmallestWidth as optional for
SplitRule Builder so that:
1. Allow the app to use the library default value if not set.
2. Allow us to introduce more bounds related options in the future.

Relnote: "
1. Take min dimensions in DP instead of pixels for SplitRule.
2. Refactor for SplitRule Builder to take min dimensions as optional.
"
Bug: 229656253
Test: pass existing
Merged-In: I95f17ce89844d18708e6fb03924d38a00af115ac
Change-Id: I95f17ce89844d18708e6fb03924d38a00af115ac