Unhide the clearCredentialSession api.
This API is independent of the get-credential api because at
the time of the get call, information about whether the previous /
existing user state should be cleared has already been lost. That is,
we can't have this as a parameter to the get call because it'd be
impossible to set it at that point.

Relnote: "Added a CredentialManager clearCredentialSession api"
Bug: 247373890
Test: CredentialManager*Test
Change-Id: Iff324bfae66c36dbc137a17b3a7d0af91d07c3bb