[CCT] Make CustomTabsCallback APIs asynchronous
In order to address the jank introduced from CustomTabs callbacks,
turns most of the APIs asynchronous. This will let us avoid being
blocked on a response from the clients.

The exception is |postMessage|-related APIs for which the calling
order needs to be guaranteed.

Test: ./gradlew browser:browser:assemble && ./gradlew
      browser:browser:connectedAndroidTest
Relnote: Make parts of CustomTabsCallback APIs asynchronous.
Bug: crbug.com/1330324
Change-Id: Ic86df723f86dce0619a0bb0bdfb9ec6111c56711