Migrate SQLiteOpenHelperWriter and ValidationWriter impls to XPoet.
This migrations allow Room to generate the RoomOpenHelper.Delegate in Kotlin and moves it back to be an inner class of the database impl, instead of a separate class.

This CL in part reverts aosp/2290293.

Also, renamed the param name of most RoomOpenHelper.Delegate functions from 'database' to 'db', aligning with the other functions, even though this is regarded as an incompatible change, these functions are only called from generated code that up until now was Java code, thus it is a safe change.

Bug: 127483380
Test: DatabaseKotlinCodeGenTest + existing
Relnote: Revert the API changes in aosp/2290293, but also renamed the db param name of most RoomOpenHelper.Delegate.
Change-Id: I1b11d32c24d1489b8f03a8fc723252e8501f2fa7