[CameraPipe] Sync fix for correct display selection based on state
DisplayInfoManager.getMaxDisplay incorrectly choses the display when multiple displays are present. It can end up choosing a display in off state in current implementation.

To fix it, logic from aosp/1974040 and aosp/2267063 are sync'ed.
The method will return the max size display that is not in off state. It will return max size off state display as a fallback if all displays are in off state.

Bug: b/255754025
Test: ./gradlew bOS
Change-Id: I13013637def795d7f5c5356b5b4a8ea8ba0b871b
Signed-off-by: Tahsin Masrur <tahsinm@google.com>