Re-implement ConstraintSet DSL to be immutable
This makes it so that ConstraintSets are not evaluated as lambdas at
measure time, instead, their parameters are packaged the same way as
a parsed Json ConstraintSet, so that we don't have to reimplement the
logic to apply those parameters and so that ConstraintSets are properly
immutable.

As a result, the ConstraintLayout Composable can now properly recognize
when the same ConstraintSet is provided at recomposition. Additionally,
the expectation that a remembered ConstriantSet does not change is now
respected.

Relnote: N/A

Bug: 258733463
Test: ConstraintLayoutTest, ConstraintSetStabilityTest
Change-Id: I4e71920d8732d105ca6703f894b0410b7471e736