Code generation for Lists
AIDL generated code uses Array<T> for non-primitive types, but uses specialized primitive arrays for primitives (e.g. IntArray for ints, CharArray for chars, etc.). These specialized arrays are disjoint from Array in the class hierarchy, so they need to be used explicitly. This adds a bit of extra complexity in the code generation. Also, shorts are not supported so we use int arrays for them.

Nested lists are not supported, but we can have Lists of Values with Lists inside. We can abstract this more seamlessly in the future.

Test: ./gradlew :privacysandbox:tools:tools-{apigenerator,apicompiler,apipackager,core}:test
Bug: 259568463
Change-Id: Iecc07fe73336cc4a5a450d066667e2b14e74540b