Fix up test issues for ScalingLazyLayoutInfo on XL Device types
Due to the different density on some screen types some of the ScalingLazyListLayoutInfoTest were not creating large enough viewports for their test contents.
To correct take into account the gaps between items and also ensure that the viewport is large enough to ensure that their is room to place the autoCentered item (even if it not the 0th item).

Bug: 260080303
Test: ScalingLazyListLayoutInfoTest
Change-Id: I941d5cd6a0ea12792ed9fe0136794a38f7afbd9e