Fix test UiObject2Test#testSetGestureMargins.
Cause of failure may be that the testing device's screen is so large
that the margins inside the pinch area form a valid rectangle that could
perform a pinch action (originally the set margins should not form a
rectangle, and the pinch action should not perform as a result).

This CL adjusts the margins to be the double of the screen size, which
should prevent forming a valid rectangle.

Also updated UiObject2Test#testSetGestureMargin.

Bug: 260235822
Test: Run this test on Pixel 2 API 26, 27, 28 and Nexus 5 API 21, and
check whether it passes.

Change-Id: Ib30594cdc858f6d0f716b72952eb4c248c51b268