ScalingLazyColumn - Estimate autoCentering topPadding when the contents are empty
When we have a ScalingLazyColumn that is created with no list items we currently do not calculate the size of the top autocentering padding. As a result when list items are then added there is a frame or two when the added items will appear at the top of the screen while the top Spacers size is calculated with a derived state calculation. This results in a frame or two where the content appears in the wrong place until to spacer is inflated and the contents can now be drawn at the center of the screen. This change calculates the amount of top padding needed by assuming an initial item of height 0.dp. For ScalingLazyListAnchorType.ItemStart this will calculate the correct top padding, for ScalingLazyListAnchorType.ItemCenter this calculation will be incorrect as the height of the items is needed to correctly size the contents.

Bug: 260406822
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Change-Id: I239a415ab2e2427c94d4c23ee74711095b53c309