Rename SettableSurface to SurfaceEdge and change the inheritance to composition.
For stream sharing, a node can have multiple downstream UseCases. When a downstream UseCase resets, it might provide a different Surface. To handle this, the edge needs to replace the DeferrableSurface. Currently, SettableSurface inherits DeferrableSurface which is irreplaceable.

- Rename SettableSurface to SurfaceEdge.
- Create an innner class "SettableSurface". It contains the logic for setting another DeferrableSurface as the source.
- Make the SurfaceEdge no longer inherit DeferrableSurface. Instead, it's composed of a SettableSurface and exposes it via a getter.

Next: add a SurfaceEdge#clearProvider call so that the connection can be reset.

Bug: 259308680
Test: manual test and ./gradlew bOS
Change-Id: I5cfe9ebfd86f2be331fc696561ed6ea4ae73f61d