Fix a11y for OutlinedTextField with label and placeholder
For a long placeholder, the label of an OTF can get completely
covered by the (transparent) placeholder in the unfocused state.
This prevents Talkback from being able to focus on the text field.

This CL repositions the placeholder so its top is always strictly
below the top of the label.

Fixes: b/251162419
Test: Added
Change-Id: Iec12140093c6c263074c679bacf7361ab3a91b17