Remove ModalBottomSheetLayout expanded anchor when sheet empty
When the sheet is empty, we will now only have a Hidden anchor. This enables passing in empty sheet content if desired by the user.

Relnote: Fixed an issue where ModalBottomSheetLayout would crash if the sheet content was empty. ModalBottomSheetLayout now allows empty sheet content. If the sheet content is empty, it will only have a Hidden state.
Test: modalBottomSheet_anchorsChange_retainsCurrentValue, modalBottomSheet_emptySheet_expandDoesNotAnimate
Bug: b/200980998, b/216693030
Change-Id: Ic2288017e3c5a80a0951cf136064cca5e2fab7cb