Fix provideContent to work with multiple active widgets
This change fixes provideContent by moving the receiver for the content
into the coroutineContext, instead of it being a member of the
GlanceAppWidget instance. There is only one GlanceAppWidget instance
that may potentially service multiple widgets, so using the same
receiver can cause issues.

Also, SessionWorker is updated so that it ignores spurious wakeups of
the Recomposer. This happens with multiple active Sessions; state
changes in one Composition causes all Recomposers to wake up. Now, only
Sessions whose Recomposer's change count has increased will receive
the event.

Simplified Session.provideGlance to return a @Composable instead of a
flow. The contentFlow is now just in GlanceAppWidget.runGlance, and is
collected in Session.provideGlance.

Fixed lambda actions to work with compound buttons.

This reverts commit 638f10f0aa771f421ccad37c71d8e7d8ed15dfe3 as well.
Tests started failing due to the broken provideContent, but they should
start working again. Includes a fix for flaky LazyColumnTest as well.

Bug: 261993523
Bug: 239747024
Test: :glance:glance-appwidget:{test,connectedCheck}
Change-Id: Idb7502d93563633922958e8c402bf1c57f091f71