Fix broken kdoc link in Color.takeOrElse.
In IntelliJ/Studio, two `[code] [links]` side by side are parsed as a
single link with a label, which means the second one is taken as the
target of the link and not rendered as text. Since this is an extension
method on Color, I don't think it helps much to link to the Color class
itself in this case, so I just removed the first link.

Also fixed some auto-formatting stuff.

Test: n/a
Change-Id: I71723f06b6ecea72a72cf3eacd2ed2c89180c60e