Replace `Closeable` with `AutoCloseable` in most APIs
This is a general cleanup CL to replace most interfaces that
previously used `Closeable` with `AutoCloseable`, the parent
interface. The primary difference is that `AutoCloseable` declares
that it might throw a general purpose exception, where Closeable
declares an io specific exception.

This should reduce some of the spurious warnings messages about
closing things from the main thread, which is normally safe.

Test: ./gradlew\
 :camera:camera-camera2-pipe:testDebugUnitTest\
 :camera:camera-camera2-pipe-testing:testDebugUnitTest\
 :camera:camera-camera2-pipe-integration:testDebugUnitTest

Change-Id: Icc6dc63a323b961db3dcc0d4bcc17f25ade36405