Fix interaction between clearComplicationData and upper layer
Recent changes to complication handling introduced a StateFlow and
better equality operator for ComplicationData. This means
ComplicationSlotsManager#clearComplicationData which operates at a
level below this, was now dangerous since it would be sticky
despite WCS sending an updated complication when switching between
favourites (because from the StateFlow's point of view nothing
changed).

Bug: 264414533

Test: Manual testing plus new unit test
Change-Id: Iadc2f5e68283aca45504ce3a49d76169251ae83b