Fix failing input mode related tests on API 34
We currently have two touch mode related changes in API34:
1. Each window gets its own Input Mode, and the unlike
previous versions, the default input mode is Keyboard
mode and not Touch mode.
2. The touch mode change listener is not called for all windows
 views, but is only called for views in windows which have focus.

The code in compose sets the touchmode during initialization. We
now initialize the touch mode during onAttachToWindow. Also,
InstrumentationRegistry sets the global touch mode state, and we
need to set the touch mode for the window hosting ComposeView. So
instead of using InstrumentationRegistry we set the touch mode using
The Compose InputManager API. We also observed some flakiness when
touch mode was changed, so we now clean up after every test by
resetting the touch mode to the default value.

Fixes: 262903257
Fixes: 262909049
Fixes: 262329840
Fixes: 262329591
Fixes: 262329583
Bug: 266755180
Bug: 267253920
Test: ./gradlew compose:ui:ui:cC -P android.testInstrumentationRunnerArguments.package=androidx.compose.ui.input.InputModeTest
Test: ./gradlew compose:f:f:cC -P android.testInstrumentationRunnerArguments.package=androidx.compose.foundation.ClickableTest
Test: ./gradlew compose:f:f:cC -P android.testInstrumentationRunnerArguments.package=androidx.compose.foundation.selection.SelectableTest
Test: ./gradlew compose:f:f:cC -P android.testInstrumentationRunnerArguments.package=androidx.compose.foundation.selection.ToggleableTest
Change-Id: Ie0984d6373989ae4c92c216200ce5e5d4b3caae0