Refactor CameraUseCaseAdapter to separate camera UseCases from app UseCases.
Previously, a UseCase represents a stream connects the camera and the app. The assumption is no longer true with Extensions, where for one app UseCases, there might be two camera UseCases. With stream sharing, it's also possible create one camera UseCase to wrap around two app UseCases.

To better reflect this reality, we update the logic in CameraUseCaseAdapter. Instead of tracking the UseCase in a single place, we track the UseCases separately: UseCases connected to the camera and UseCases connected to the app.

Changes:
- Track the Extentions placeholder UseCase with a single variable instead of a list. The extra UseCase can only be one at most.
- Added a mCameraUseCases that tracks all the UseCases connected to the camera.
- For both addUseCases() and removeUseCases(), have them to go through the same process. The process will 1) calculate the new states, 2) calculate the action items to go to the new states, 3) take actions and 4) track the new states if the actions are successful.

This change is covered by existing tests.

Bug: 265531423
Test: manual test and ./gradlew bOS
Change-Id: I69e0c3dd9259b9f9400ac3c12468ad49b48340fe