Marking PivotOffsets as @Immutable
PivotOffsets is a class that is intended to hold the parent and
child offset values and are not updatable once created.

The places where PivotOffset was being used as a default value for
parameters was marked as dynamic by the compose-compiler.

To unlock some of the compiler optimizations, marking PivotOffsets
as immutable to allow the default values to be interpreted as
static by the compiler.
Source: https://github.com/androidx/androidx/blob/androidx-main/compose/compiler/design/compiler-metrics.md#default-parameter-expressions-that-are-dynamic

Test: NA
Relnote: "Marking PivotOffsets as @Immutable"

Change-Id: I8385b11ad151f8814adf8c92a22f39447d835bfb