Adds setLargeIcon(Icon) API to NotifCompat.Builder
Adds setLargeIcon(Icon) API to NotificationCompat.Builder.
Also makes public the previously private function that scales Bitmaps to
the appropriate size for use as LargeIcons; this is required only for
API versions 23 to 27, but cannot be done once a bitmap is wrapped in an
Icon (as the underlying Bitmap is private).

Also added a unit test covering the scaling code in
setLargeIcon(Bitmap) since it was previously untested.

A previous version of this code was submitted with a bad test that would
inappropriately test for an extra being set when a null icon was set.
Prior to API version 24, EXTRA_LARGE_ICON was not set if largeIcon was
set to null. Starting in version 24, EXTRA_LARGE_ICON is set, but its
value is null. The tests have been adjusted accordingly.

Bug: 172282791
Test: Unit tests (now including testing on API version 19-21)
Relnote: Adds NotifCompat.Builder.setLargeIcon(Icon) signature
Change-Id: Ic61727654e961597ef0c75cb75cd0446bb1a09eb