Handle absence of previous target in ModalBottomSheetAnchorChangeHandler
When we animate or snap, we update the currentValue regardless of whether it exists in the anchors. If ModalBottomSheetState was moved to a visible value (HalfExpanded or Expanded) while only having a hidden anchor (when the sheet content is empty), we would crash in the AnchorChangeHandler because we required the previous target to be present in the previous set of anchors.

Test: modalBottomSheet_shortSheet_anchorChangeHandler_previousTargetNotInAnchors_reconciles, modalBottomSheet_tallSheet_anchorChangeHandler_previousTargetNotInAnchors_reconciles
Relnote: Fixed a bug in ModalBottomSheetLayout where the sheet would crash when going from the hidden to a visible state in some circumstances.
Bug: 265444789
Change-Id: Ia926514a65dc4d2c44781e132f8faae7e442cf68