Add allowlist for interface inheritance.
The error message should leave out allowlisted classes, but this is
difficult to test without adding stubs support to the
InterfaceParserTest - I might do this in a future change.

I've added an allowlisted inheritance case to FullFeaturedSdkTest so
accordingly updated the goldens - of course, these will change when the
updated code gen is implemented.

Test: ./gradlew :privacysandbox:tools:tools-{apigenerator,apicompiler,apipackager,core}:test --parallel
Bug: 265266769
Change-Id: I694ba272fdc51f360bccac9b7b5215447cb5bdfd