Move the FragmentNavigator to transition APIs
Moving the FragmentNavigator to use the NavigatorState transition APIs.
This means that NavBackStackEntries will now honor the lifecycle state
of Fragments.

This also moves the system back handling behavior away from the
NavController to the FragmentNavigator itself.

RelNote: "When using Navigation with Fragments, the NavBackStackEntries
associated with the fragment destinations will now respect the Lifecycle
of the Fragment."
Test: adding tests
Bug: 238686802

Change-Id: I3cb193646619c371ce4024b19d709fbbfada32be