Rework dialog transitions to match UX motion spec more closely
The intro transition was more rapid than specified, whereas the outro was smooth. The code is now symmetrical between the intro/outro cases, both of which are controlled by the 'showDialog' boolean.

Bug: 267285621
Relnote: "We have improved the dialog transitions to match the UX motion spec - the intro transition is now smoother to match the outro transition."
Test: Manually tested
Change-Id: Ib5af9b7c468543e8fb5988b50bf712bbfb1a340f