Make the open Health Connect Action public.
This intent has been used by the HC app for the past year, but we are
adding it to the SDK now so that we can have a different value for
Android U, without requiring developers to handle this on their side.

RelNote: Adding an intent to use for opening Health Connect.
Bug: 266069579
Test: N/A
Change-Id: Ic80557de25d163b45f1ea2bdd54efb9d6079a704