Fix horizontal insets for multi-row TopAppBar
Applies the horizontal portion of the window insets to the second row of
the TwoRowsTopAppBar (used by MediumTopAppBar and LargeTopAppBar)

Previously, the window insets were only being applied to the first row,
which resulted in the second row content being cutoff by the insets when
there were horizontal insets (such as a phone in landscape mode with
3-button navigation).

Now, the horizontal portion of the insets is applied to the second row
as well, to avoid being cutoff. (Note: we don't want to apply the top insets
again, since that would result in insets being doubled applied by both
rows).

Fix: 267647327
Test: Manual with 3 button navigation

Change-Id: Ib1d90564f8942c372808623ff031cb1dd9b02609