NavHost compose preview to show startDestination
When previewing a composable with NavHost, it will now show the NavGraph's startDestinatin by default. Any dialogs that are shown by default in the startDestination will also show in the preview.

Test: ./gradlew navigation:navigation-compose:cC
Bug: 253654317
Relnote: "When previewing a composable with NavHost, it will now show the NavGraph's startDestinatin by default. Any dialogs that are shown by default in the startDestination will also show in the preview."
Change-Id: I2b89f0e3957f9727912b94e20efdfb6b115b97e9