Update child rotation degrees when the stream is mirroed by StreamSharing.
In CameraX, the transformation info given to the app includes the crop rect, the rotation degrees, and (inexplicitly) the mirroring. They should be applied in the order of crop -> rotation -> mirroring. The StreamSharing applies cropping and mirroring to the buffer, but not the rotation. Once the stream is mirrored, the rotation also needs to be updated to reflect that.

- Overload UseCase#getRelativeRotation and accept a param of whether the output require mirroring.
- In Preview, use the new method to calculate rotation.

Bug: 265818567
Test: manual test and ./gradlew bOS
Change-Id: I49e342e03b1e0ed4c874d84a7d88513b35ddb659