Make the Photo Picker fallback action and extras public
When MediaStore.ACTION_PICK_IMAGES is unavailable,
the PickVisualMedia and PickMultipleVisualMedia contracts
are responsible for the fallback behavior.

In order to provide a level playing field to all devices
built on top of AOSP, the constants for handling this
fallback path should be documented as public constants
that OEMs or other system apps can use to implement their
own Photo Picker.

This replaces the previous internal only constants
that will be removed once their users have swapped to
the new action and extra.

Relnote: "The action and extra used by the
`PickVisualMedia` and `PickMultipleVisualMedia` contracts
as a fallback for when `MediaStore.ACTION_PICK_IMAGES` is
unavailable are now public constants that provide
API stability to OEMs and system apps that want to
provide a consistent Photo Picker experience. The
implementation of this fallback is still limited to only
system apps."
Test: ./gradlew checkApi
BUG: 264622388

Change-Id: Icd320c81c44c718144c55fcd7c5e39056b3bdc7f